Ajout IdRight dans Instant Gestion des droits dans checkData pour filtrer les ressources a renvoyer Ajout boostrap et images en local Rennomage méthode instant, moment, album pour une meilleure compréhension

+++ src/main/java/pfe/ece/LinkUS/Service/NotificationTokenServiceImpl.java
@@ -32,7 +32,9 @@ public class NotificationTokenServiceImpl implements NotificationTokenService{
     public void removeNotifcationTokenByUsername(String username) {
 
         NotificationToken notificationToken = notificationTokenRepository.findByUsername(username);
-        notificationTokenRepository.delete(notificationToken);
+        if(notificationToken != null) {
+            notificationTokenRepository.delete(notificationToken);
+        }
     }
 
     @Override
